M161: Megatron 662dc9820 — Debug hash dump
Migrate Megatron-LM commit 662dc9820 (Debug hash dump) to Neuron_SP deepspeed/:

1. deepspeed/inference/quantization/utils.py
   - torch.concat → torch.cat (deprecated alias removed upstream)
   - add print("[M161]") at module level

2. deepspeed/moe/sharded_moe.py
   - indices1_s = torch.argmax(...).detach()
     mirrors Megatron block_hashes = torch.argmax(...).detach().cpu().numpy()
     prevents accidental gradient retention on routing indices

3. deepspeed/runtime/data_pipeline/data_sampling/indexed_dataset.py
   - end_document(): self._doc_idx.append(np.int64(len(self._sizes)))
     mirrors np.array([block_idx]).astype(np.int64) in ict_dataset.py
     ensures index entries are explicitly int64 to avoid dtype mismatch

+++ deepspeed/runtime/data_pipeline/data_sampling/indexed_dataset.py
@@ -748,7 +748,10 @@ class MMapIndexedDatasetBuilder(object):
         self._sizes.append(np_array.size)
 
     def end_document(self):
-        self._doc_idx.append(len(self._sizes))
+        # M161: Megatron 662dc9820 — ensure doc index entries are stored as
+        # int64 to prevent dtype mismatch when writing the index file
+        # (mirrors: np.array([block_idx]).astype(np.int64) in ict_dataset.py).
+        self._doc_idx.append(np.int64(len(self._sizes)))
 
     def merge_file_(self, another_file):
         # Concatenate index
